resync

+++ src/com/mcnsa/chat/type/ChatPlayer.java
@@ -21,7 +21,7 @@ public class ChatPlayer implements Serializable{
 	public ArrayList<String> muted = new ArrayList<String>();
 	public ArrayList<String> serversVisited = new ArrayList<String>();
 	transient Players playersFile;
-	public int timeoutTill = 0;
+	public long timeoutTill = 0;
 	
 	@SuppressWarnings("unchecked")
 	public ChatPlayer(String username){
@@ -63,7 +63,7 @@ public class ChatPlayer implements Serializable{
 			this.modes.put("POOF", this.playersFile.get().getBoolean("modes.POOF"));
 			this.muted = (ArrayList<String>) this.playersFile.get().getList("muted");
 			this.serversVisited = (ArrayList<String>) this.playersFile.get().getList("serversVisited");
-			this.timeoutTill = this.playersFile.get().getInt("timeoutTill");
+			this.timeoutTill = this.playersFile.get().getLong("timeoutTill");
 		}
 		playersFile.save();
 	}
